Updated names of auth guards for clarity
