--- prime-angular-frontend/src/app/modules/auth/shared/guards/auth-redirect.guard.spec.ts
@@ -1,15 +1,0 @@
-import { TestBed, async, inject } from '@angular/core/testing';
-
-import { AuthRedirectGuard } from './auth-redirect.guard';
-
-describe('AuthRedirectGuard', () => {
-  beforeEach(() => {
-    TestBed.configureTestingModule({
-      providers: [AuthRedirectGuard]
-    });
-  });
-
-  it('should ...', inject([AuthRedirectGuard], (guard: AuthRedirectGuard) => {
-    expect(guard).toBeTruthy();
-  }));
-});
