+++ prime-angular-frontend/src/app/modules/auth/shared/guards/authentication.guard.spec.ts
@@ -1,0 +1,15 @@
+import { TestBed, async, inject } from '@angular/core/testing';
+
+import { AuthenticateGuard } from './authentication.guard';
+
+describe('AuthenticateGuard', () => {
+  beforeEach(() => {
+    TestBed.configureTestingModule({
+      providers: [AuthenticateGuard]
+    });
+  });
+
+  it('should ...', inject([AuthenticateGuard], (guard: AuthenticateGuard) => {
+    expect(guard).toBeTruthy();
+  }));
+});
