+++ prime-angular-frontend/src/app/modules/auth/shared/guards/authorization-redirect.guard.ts
@@ -6,22 +6,20 @@ import {
 
 import { Observable } from 'rxjs';
 
-import { KeycloakLoginOptions } from 'keycloak-js';
 import { KeycloakAuthGuard, KeycloakService } from 'keycloak-angular';
 
 import { APP_CONFIG, AppConfig } from 'app/app-config.module';
+import { Role } from '@auth/shared/enum/role.enum';
 import { LoggerService } from '@core/services/logger.service';
-import { AuthService } from '@auth/shared/services/auth.service';
 
 @Injectable({
   providedIn: 'root'
 })
-export class AuthGuard extends KeycloakAuthGuard implements CanActivateChild {
+export class AuthorizationRedirectGuard extends KeycloakAuthGuard implements CanActivateChild {
   constructor(
     protected router: Router,
     protected keycloakAngular: KeycloakService,
     @Inject(APP_CONFIG) private config: AppConfig,
-    private authService: AuthService,
     private logger: LoggerService
   ) {
     super(router, keycloakAngular);
@@ -42,21 +40,23 @@ export class AuthGuard extends KeycloakAuthGuard implements CanActivateChild {
   public isAccessAllowed(route: ActivatedRouteSnapshot, state: RouterStateSnapshot): Promise<boolean> {
     return new Promise((resolve, reject) => {
       if (!this.authenticated) {
-        // Capture the user's current location, and provide it to
-        // Keycloak to redirect the user to where they originated
-        // once authenticated
-        const options: KeycloakLoginOptions = {
-          redirectUri: state.url
-        };
+        return resolve(true);
+      }
 
-        this.keycloakAngular.login(options)
-          .catch(e => this.logger.error(e));
-
+      if (this.keycloakAngular.isUserInRole(Role.ENROLLEE)) {
+        this.router.navigate([this.config.routes.enrolment]);
+        return reject(false);
+      } else if (
+        this.keycloakAngular.isUserInRole(Role.PROVISIONER) ||
+        this.keycloakAngular.isUserInRole(Role.ADMIN)
+      ) {
+        this.router.navigate([this.config.routes.provision]);
         return reject(false);
       }
 
-      // Otherwise, allow current route access
-      return resolve(true);
+      // Access has been denied
+      this.router.navigate([this.config.routes.denied]);
+      return reject(false);
     });
   }
 }
