+++ prime-angular-frontend/src/app/modules/enrolment/enrolment-routing.module.ts
@@ -3,7 +3,7 @@ import { Routes, RouterModule } from '@angular/router';
 
 import { CanDeactivateFormGuard } from '@core/guards/can-deactivate-form.guard';
 import { DashboardComponent } from '@shared/components/dashboard/dashboard.component';
-import { AuthGuard } from '@auth/shared/guards/auth.guard';
+import { AuthenticateGuard } from '@auth/shared/guards/authentication.guard';
 
 import { EnrolleeGuard } from './shared/guards/enrollee.guard';
 import { EnrolmentGuard } from './shared/guards/enrolment.guard';
@@ -24,7 +24,7 @@ const routes: Routes = [
     // Check authentication and authorization each time
     // the router navigates to the next route
     canActivateChild: [
-      AuthGuard,
+      AuthenticateGuard,
       // Guard module from being accessed without the proper
       // authorization based on the user role permissions
       EnrolleeGuard
